+++ src/components/atoms/BoulderButtons.svelte
@@ -1,15 +1,16 @@
 <script lang="ts">
-	import { clickedCells, selector } from '../../stores/BoulderStore.svelte';
+	import { clickedGrips, selector } from '../../stores/BoulderStore.svelte';
 	import Button from './Button.svelte';
 
 	export const variant: 'default' | 'play' = 'default';
 	export let handleSaveBoulder: () => void;
+	export let handleDisplayBoulder: () => void;
 </script>
 
 {#if variant === 'default'}
 	<div class="grid w-[20.8em] grid-flow-col justify-stretch gap-4 pl-9 pr-1 pt-4 sm:w-[23.5em]">
 		<Button variant="outline" on:click={() => selector.setMode('Start')}>Start</Button>
-		<Button variant="outline" on:click={() => selector.setMode('Top')}>Top</Button>
+		<Button variant="outline" on:click={() => selector.setMode('Finish')}>Finish</Button>
 		<Button emoji="💾" variant="outlineGreen" aria-label="Save" on:click={handleSaveBoulder}
 		></Button>
 		<Button
@@ -17,8 +18,16 @@
 			variant="outlineYellow"
 			aria-label="Clear"
 			on:click={() => {
-				clickedCells.clear();
+				clickedGrips.clear();
 				selector.clear();
+			}}
+		></Button>
+
+		<Button
+			emoji="▶"
+			variant="outline"
+			on:click={() => {
+				handleDisplayBoulder();
 			}}
 		></Button>
 	</div>