+++ src/components/utils/BoulderTypes.ts
@@ -1,21 +1,29 @@
-export type Mode = 'Top' | 'Start';
+export type Mode = 'Finish' | 'Start';
 
-export type Cell = {
+export type Grip = {
 	id: string;
 	colorBrightness: string;
+	start?: string;
+	finish?: string;
 };
 
 export type Selector = {
 	selectedMode: Mode | undefined;
-	selectedStartCell: string | undefined;
-	selectedTopCell: string | undefined;
+	selectedStartGrip: string | undefined;
+	selectedFinishGrip: string | undefined;
 };
 
 export type Boulder = {
 	id: string;
-	createdAt: number;
+	createdAt: string;
 	name?: string;
 	start?: string;
-	top?: string;
-	path: Cell[];
+	finish?: string;
+	path: Grip[];
+};
+
+export type ApiResponse = {
+	status: number;
+	title: string;
+	detail: string;
 };